pyproject.toml
[project]
name = "conversational-agent"
version = "0.0.1"
dependencies = [
    "google-genai==1.5.0",
    "chainlit==2.2.1",
    "pydantic-settings==2.8.1",
    "python-json-logger==3.3.0",
    "openpyxl==3.1.5",
    "gspread-dataframe==4.0.0",
    "spacy==3.8.3",
    "fpdf==1.7.2",
    "matplotlib==3.9.4",
]